+++ examples/hello.ino
@@ -16,7 +16,7 @@
 
 uint16_t counter;
 
-const size_t serviceDataSize = 1;
+const microapp_size_t serviceDataSize = 1;
 const uint16_t serviceDataUuid = 0x1234;
 uint8_t serviceData[serviceDataSize];
