+++ src/ServiceData.cpp
@@ -4,7 +4,7 @@ void ServiceData_::write(uint16_t appUuid, const char* str) {
 	uint8_t* payload                         = getOutgoingMessagePayload();
 	microapp_sdk_service_data_t* serviceData = reinterpret_cast<microapp_sdk_service_data_t*>(payload);
 	serviceData->appUuid                     = appUuid;
-	size_t size                              = strlen(str);
+	microapp_size_t size                              = strlen(str);
 	// Silently truncate data if too long
 	if (size > MICROAPP_SDK_MAX_SERVICE_DATA_LENGTH) {
 		serviceData->size = MICROAPP_SDK_MAX_SERVICE_DATA_LENGTH;
@@ -31,7 +31,7 @@ void ServiceData_::write(uint16_t appUuid, String str) {
 	_write(serviceData);
 }
 
-void ServiceData_::write(uint16_t appUuid, uint8_t* buf, size_t size) {
+void ServiceData_::write(uint16_t appUuid, uint8_t* buf, microapp_size_t size) {
 	uint8_t* payload                         = getOutgoingMessagePayload();
 	microapp_sdk_service_data_t* serviceData = reinterpret_cast<microapp_sdk_service_data_t*>(payload);
 	serviceData->appUuid                     = appUuid;
